backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
langchain>=0.0.200
pinecone-client>=2.2.1
openai>=0.27.0
python-dotenv>=0.19.0
tiktoken>=0.3.3
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
nltk>=3.6.3
spacy>=3.1.0
redis>=4.3.4
aioredis>=2.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
requests>=2.26.0
aiohttp>=3.8.1
asyncio>=3.4.3
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.24.0 